fix: address new Claude Code review concerns - uint64 consistency and health monitoring
Addresses the latest Claude Code review feedback on PR #35.

Changes:
1. Complete uint64 Migration (eliminates overflow validation concern):
   - Changed calculateCapacity() parameters from int to uint64
   - Changed assignSlice() parameters from int to uint64
   - Removed all int conversions for numActivators/activatorIndex
   - Updated test structs to use uint64 throughout
   - No more gosec nolint comments needed for these conversions

2. Feature Gate Consistency:
   - Snapshot feature gates at start of processStateUpdate()
   - Pass qpAuthority as parameter to handleExistingPodEvent()
   - Prevents mid-operation ConfigMap updates from causing inconsistent behavior

3. Worker Health Monitoring:
   - Added lastPanicTime atomic field to track panic recency
   - Implemented IsWorkerHealthy() method
   - Returns false if panicked in last minute or queue > 50% full
   - Enables external health checks to detect panic loops

All tests pass. This completes all concerns from the latest review.

+++ pkg/activator/net/throttler_test.go
@@ -303,11 +303,11 @@ func TestThrottlerCalculateCapacity(t *testing.T) {
 	logger := TestLogger(t)
 	tests := []struct {
 		name                 string
-		numActivators        int32
-		containerConcurrency int
-		numTrackers          int
-		activatorCount       int
-		backendCount         int
+		numActivators        uint64
+		containerConcurrency uint64
+		numTrackers          uint64
+		activatorCount       uint64
+		backendCount         uint64
 	}{{
 		name:                 "over revisionMaxConcurrency",
 		numActivators:        200,
@@ -322,8 +322,8 @@ func TestThrottlerCalculateCapacity(t *testing.T) {
 				logger:  logger,
 				breaker: newInfiniteBreaker(logger),
 			}
-			rt.containerConcurrency.Store(uint64(tt.containerConcurrency))
-			rt.numActivators.Store(uint64(tt.numActivators))
+			rt.containerConcurrency.Store(tt.containerConcurrency)
+			rt.numActivators.Store(tt.numActivators)
 			// Default to first activator (index 1) in 1-based indexing when not specified
 			rt.activatorIndex.Store(1)
 			// shouldn't really happen since revisionMaxConcurrency is very, very large,
